refactor(commons): add parameters and key kinds to concrete-commons
