+++ concrete-commons/src/numeric/signed.rs
@@ -1,9 +1,11 @@
-use super::{CastFrom, Numeric, UnsignedInteger};
-use crate::CastInto;
 use std::ops::{
     Add, AddAssign, BitAnd, BitAndAssign, BitOr, BitOrAssign, BitXor, BitXorAssign, Div, DivAssign,
     Mul, MulAssign, Neg, Not, Rem, RemAssign, Shl, ShlAssign, Shr, ShrAssign, Sub, SubAssign,
 };
+
+use crate::numeric::CastInto;
+
+use super::{CastFrom, Numeric, UnsignedInteger};
 
 /// A trait shared by all the unsigned integer types.
 pub trait SignedInteger:
