+++ concrete-commons/src/numeric/unsigned.rs
@@ -1,9 +1,11 @@
-use super::{CastFrom, Numeric, SignedInteger};
-use crate::CastInto;
 use std::ops::{
     Add, AddAssign, BitAnd, BitAndAssign, BitOr, BitOrAssign, BitXor, BitXorAssign, Div, DivAssign,
     Mul, MulAssign, Not, Rem, RemAssign, Shl, ShlAssign, Shr, ShrAssign, Sub, SubAssign,
 };
+
+use crate::numeric::CastInto;
+
+use super::{CastFrom, Numeric, SignedInteger};
 
 /// A trait shared by all the unsigned integer types.
 pub trait UnsignedInteger:
