+++ concrete-commons/src/parameters.rs
@@ -1,0 +1,81 @@
+use serde::{Deserialize, Serialize};
+
+/// The number plaintexts in a plaintext list.
+#[derive(Copy, Clone, Eq, PartialEq, Debug, Serialize, Deserialize)]
+pub struct PlaintextCount(pub usize);
+
+/// The number messages in a messages list.
+#[derive(Copy, Clone, Eq, PartialEq, Debug, Serialize, Deserialize)]
+pub struct CleartextCount(pub usize);
+
+/// The number of ciphertexts in a ciphertext list.
+#[derive(Copy, Clone, Eq, PartialEq, Debug, Serialize, Deserialize)]
+pub struct CiphertextCount(pub usize);
+
+/// The number of scalar in an LWE mask + 1 .
+#[derive(Debug, PartialEq, Eq, Copy, Clone, Serialize, Deserialize)]
+pub struct LweSize(pub usize);
+
+impl LweSize {
+    /// Returns the associated [`LweDimension`].
+    pub fn to_lwe_dimension(&self) -> LweDimension {
+        LweDimension(self.0 - 1)
+    }
+}
+
+/// The number of scalar in an LWE mask, or the length of an LWE secret key.
+#[derive(Copy, Clone, Eq, PartialEq, Debug, Serialize, Deserialize)]
+pub struct LweDimension(pub usize);
+
+impl LweDimension {
+    /// Returns the associated [`LweSize`].
+    pub fn to_lwe_size(&self) -> LweSize {
+        LweSize(self.0 + 1)
+    }
+}
+
+/// The number of polynomials of an GLWE mask + 1.
+#[derive(Debug, PartialEq, Eq, Copy, Clone, Serialize, Deserialize)]
+pub struct GlweSize(pub usize);
+
+impl GlweSize {
+    /// Returns the associated [`GlweDimension`].
+    pub fn to_glwe_dimension(&self) -> GlweDimension {
+        GlweDimension(self.0 - 1)
+    }
+}
+
+/// The number of polynomials of an GLWE mask, or the size of an GLWE secret key.
+#[derive(Copy, Clone, Eq, PartialEq, Debug, Serialize, Deserialize)]
+pub struct GlweDimension(pub usize);
+
+impl GlweDimension {
+    /// Returns the associated [`GlweSize`].
+    pub fn to_glwe_size(&self) -> GlweSize {
+        GlweSize(self.0 + 1)
+    }
+}
+/// The number of coefficients of a polynomial.
+///
+/// Assuming a polynomial $a_0 + a_1X + /dots + a_nX^N$, this returns $N+1$.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub struct PolynomialSize(pub usize);
+
+/// The number of polynomials in a polynomial list.
+///
+/// Assuming a polynomial list, this return the number of polynomials.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct PolynomialCount(pub usize);
+
+/// The logarithm of the base used in a decomposition.
+///
+/// When decomposing an integer over powers of the $2^B$ basis, this type represents the $B$ value.
+#[derive(Debug, PartialEq, Eq, Copy, Clone, Deserialize, Serialize)]
+pub struct DecompositionBaseLog(pub usize);
+
+/// The number of levels used in a decomposition.
+///
+/// When decomposing an integer over the $l$ largest powers of the basis, this type represents
+/// the $l$ value.
+#[derive(Debug, PartialEq, Eq, Copy, Clone, Deserialize, Serialize)]
+pub struct DecompositionLevelCount(pub usize);
