+++ concrete-core/benches/keyswitch.rs
@@ -1,6 +1,5 @@
+use concrete_commons::parameters::{DecompositionBaseLog, DecompositionLevelCount, LweDimension};
 use concrete_core::crypto::lwe::{LweCiphertext, LweKeyswitchKey};
-use concrete_core::crypto::LweDimension;
-use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
 use concrete_core::math::random::{RandomGenerable, UniformMsb};
 use concrete_core::math::torus::UnsignedTorus;
 use criterion::{BenchmarkId, Criterion};
