+++ concrete-core/src/crypto/bootstrap/fourier/mod.rs
@@ -1,7 +1,6 @@
 use std::cell::RefCell;
 use std::fmt::Debug;
 
-use concrete_commons::{CastInto, Numeric};
 use concrete_fftw::array::AlignedVec;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
@@ -12,13 +11,16 @@ use crate::crypto::bootstrap::Bootstrap;
 use crate::crypto::ggsw::GgswCiphertext;
 use crate::crypto::glwe::GlweCiphertext;
 use crate::crypto::lwe::LweCiphertext;
-use crate::crypto::{GlweSize, LweDimension};
-use crate::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount, SignedDecomposer};
+use crate::math::decomposition::SignedDecomposer;
 use crate::math::fft::{Complex64, Fft, FourierPolynomial};
-use crate::math::polynomial::{MonomialDegree, Polynomial, PolynomialList, PolynomialSize};
+use crate::math::polynomial::{MonomialDegree, Polynomial, PolynomialList};
 use crate::math::tensor::{AsMutSlice, AsMutTensor, AsRefSlice, AsRefTensor, IntoTensor, Tensor};
 use crate::math::torus::UnsignedTorus;
 use crate::{ck_dim_div, ck_dim_eq, zip, zip_args};
+use concrete_commons::numeric::{CastInto, Numeric};
+use concrete_commons::parameters::{
+    DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+};
 use std::marker::PhantomData;
 
 #[cfg(test)]
@@ -63,11 +65,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::FourierBootstrapKey;
-    /// use concrete_core::crypto::{GlweSize, LweDimension, LweSize};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let bsk: FourierBootstrapKey<_, u32> = FourierBootstrapKey::allocate(
     ///     Complex64::new(9., 8.),
     ///     GlweSize(7),
@@ -134,11 +136,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::FourierBootstrapKey;
-    /// use concrete_core::crypto::{GlweSize, LweDimension};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let vector = vec![Complex64::new(0., 0.); 256 * 5 * 4 * 4 * 15];
     /// let bsk: FourierBootstrapKey<_, u32> = FourierBootstrapKey::from_container(
     ///     vector.as_slice(),
@@ -205,11 +207,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::{FourierBootstrapKey, StandardBootstrapKey};
-    /// use concrete_core::crypto::{GlweSize, LweDimension, LweSize};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let bsk = StandardBootstrapKey::allocate(
     ///     9u32,
     ///     GlweSize(7),
@@ -275,11 +277,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::FourierBootstrapKey;
-    /// use concrete_core::crypto::{GlweSize, LweDimension, LweSize};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let bsk: FourierBootstrapKey<_, u32> = FourierBootstrapKey::allocate(
     ///     Complex64::new(9., 8.),
     ///     GlweSize(7),
@@ -299,11 +301,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::FourierBootstrapKey;
-    /// use concrete_core::crypto::{GlweSize, LweDimension, LweSize};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let bsk: FourierBootstrapKey<_, u32> = FourierBootstrapKey::allocate(
     ///     Complex64::new(9., 8.),
     ///     GlweSize(7),
@@ -323,11 +325,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::FourierBootstrapKey;
-    /// use concrete_core::crypto::{GlweSize, LweDimension, LweSize};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let bsk: FourierBootstrapKey<_, u32> = FourierBootstrapKey::allocate(
     ///     Complex64::new(9., 8.),
     ///     GlweSize(7),
@@ -347,11 +349,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::FourierBootstrapKey;
-    /// use concrete_core::crypto::{GlweSize, LweDimension, LweSize};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let bsk: FourierBootstrapKey<_, u32> = FourierBootstrapKey::allocate(
     ///     Complex64::new(9., 8.),
     ///     GlweSize(7),
@@ -371,11 +373,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::FourierBootstrapKey;
-    /// use concrete_core::crypto::{GlweSize, LweDimension, LweSize};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let bsk: FourierBootstrapKey<_, u32> = FourierBootstrapKey::allocate(
     ///     Complex64::new(9., 8.),
     ///     GlweSize(7),
@@ -406,11 +408,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::FourierBootstrapKey;
-    /// use concrete_core::crypto::{GlweSize, LweDimension, LweSize};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let bsk: FourierBootstrapKey<_, u32> = FourierBootstrapKey::allocate(
     ///     Complex64::new(9., 8.),
     ///     GlweSize(7),
@@ -452,11 +454,11 @@ where
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::{
+    ///     DecompositionBaseLog, DecompositionLevelCount, GlweSize, LweDimension, PolynomialSize,
+    /// };
     /// use concrete_core::crypto::bootstrap::FourierBootstrapKey;
-    /// use concrete_core::crypto::{GlweSize, LweDimension, LweSize};
-    /// use concrete_core::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
     /// use concrete_core::math::fft::Complex64;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// use concrete_core::math::tensor::{AsMutTensor, AsRefTensor};
     /// let mut bsk: FourierBootstrapKey<_, u32> = FourierBootstrapKey::allocate(
     ///     Complex64::new(9., 8.),
