+++ concrete-core/src/crypto/encoding/plaintext.rs
@@ -1,9 +1,8 @@
-use concrete_commons::Numeric;
-
-use crate::crypto::PlaintextCount;
 use crate::math::polynomial::Polynomial;
 use crate::math::tensor::{AsMutSlice, AsMutTensor, AsRefSlice, AsRefTensor, Tensor};
 use crate::{ck_dim_div, tensor_traits};
+use concrete_commons::numeric::Numeric;
+use concrete_commons::parameters::PlaintextCount;
 
 /// An plaintext (encoded) value.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
@@ -26,8 +25,8 @@ where
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PlaintextCount;
     /// use concrete_core::crypto::encoding::*;
-    /// use concrete_core::crypto::PlaintextCount;
     /// let plain_list = PlaintextList::allocate(1 as u8, PlaintextCount(100));
     /// assert_eq!(plain_list.count(), PlaintextCount(100));
     /// ```
@@ -42,8 +41,8 @@ impl<Cont> PlaintextList<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PlaintextCount;
     /// use concrete_core::crypto::encoding::*;
-    /// use concrete_core::crypto::PlaintextCount;
     /// let plain_list = PlaintextList::from_container(vec![1 as u8; 100]);
     /// assert_eq!(plain_list.count(), PlaintextCount(100));
     /// ```
@@ -62,8 +61,8 @@ impl<Cont> PlaintextList<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PlaintextCount;
     /// use concrete_core::crypto::encoding::*;
-    /// use concrete_core::crypto::PlaintextCount;
     /// let plain_list = PlaintextList::from_container(vec![1 as u8; 100]);
     /// assert_eq!(plain_list.count(), PlaintextCount(100));
     /// ```
@@ -80,7 +79,6 @@ impl<Cont> PlaintextList<Cont> {
     ///
     /// ```rust
     /// use concrete_core::crypto::encoding::*;
-    /// use concrete_core::crypto::PlaintextCount;
     /// let plain_list = PlaintextList::from_container(vec![1 as u8; 100]);
     /// plain_list.plaintext_iter().for_each(|a| assert_eq!(a.0, 1));
     /// assert_eq!(plain_list.plaintext_iter().count(), 100);
@@ -104,7 +102,6 @@ impl<Cont> PlaintextList<Cont> {
     ///
     /// ```rust
     /// use concrete_core::crypto::encoding::*;
-    /// use concrete_core::crypto::PlaintextCount;
     /// let mut plain_list = PlaintextList::from_container(vec![1 as u8; 100]);
     /// plain_list
     ///     .plaintext_iter_mut()
@@ -132,8 +129,8 @@ impl<Cont> PlaintextList<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PlaintextCount;
     /// use concrete_core::crypto::encoding::*;
-    /// use concrete_core::crypto::PlaintextCount;
     /// let mut plain_list = PlaintextList::from_container(vec![1 as u8; 100]);
     /// plain_list
     ///     .sublist_iter(PlaintextCount(10))
@@ -157,8 +154,8 @@ impl<Cont> PlaintextList<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PlaintextCount;
     /// use concrete_core::crypto::encoding::*;
-    /// use concrete_core::crypto::PlaintextCount;
     /// let mut plain_list = PlaintextList::from_container(vec![1 as u8; 100]);
     /// plain_list
     ///     .sublist_iter_mut(PlaintextCount(10))
@@ -186,9 +183,8 @@ impl<Cont> PlaintextList<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::crypto::encoding::*;
-    /// use concrete_core::crypto::PlaintextCount;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let plain_list = PlaintextList::from_container(vec![1 as u8; 100]);
     /// let plain_poly = plain_list.as_polynomial();
     /// assert_eq!(plain_poly.polynomial_size(), PolynomialSize(100));
@@ -205,9 +201,8 @@ impl<Cont> PlaintextList<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::crypto::encoding::*;
-    /// use concrete_core::crypto::PlaintextCount;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let mut plain_list = PlaintextList::from_container(vec![1 as u8; 100]);
     /// let mut plain_poly = plain_list.as_mut_polynomial();
     /// assert_eq!(plain_poly.polynomial_size(), PolynomialSize(100));
