+++ concrete-core/src/crypto/glwe/tests.rs
@@ -1,5 +1,3 @@
-use concrete_commons::LogStandardDev;
-
 use crate::crypto::encoding::PlaintextList;
 use crate::crypto::glwe::GlweList;
 use crate::crypto::secret::generators::{EncryptionRandomGenerator, SecretRandomGenerator};
@@ -8,6 +6,7 @@ use crate::math::random::RandomGenerator;
 use crate::math::torus::UnsignedTorus;
 use crate::test_tools;
 use crate::test_tools::assert_delta_std_dev;
+use concrete_commons::dispersion::LogStandardDev;
 
 fn test_glwe<T: UnsignedTorus>() {
     // random settings
