+++ concrete-core/src/crypto/lwe/tests.rs
@@ -1,21 +1,21 @@
 use concrete_npe as npe;
 
-use concrete_commons::{
-    CastFrom, DispersionParameter, LogStandardDev, Numeric, SignedInteger, Variance,
-};
-
 use crate::crypto::encoding::{Cleartext, CleartextList, Plaintext, PlaintextList};
 use crate::crypto::lwe::{LweCiphertext, LweKeyswitchKey, LweList};
 use crate::crypto::secret::generators::{EncryptionRandomGenerator, SecretRandomGenerator};
 use crate::crypto::secret::LweSecretKey;
-use crate::crypto::{CiphertextCount, CleartextCount, LweDimension, PlaintextCount};
-use crate::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
 use crate::math::random::{RandomGenerable, RandomGenerator, UniformMsb};
 use crate::math::tensor::{AsMutTensor, AsRefTensor, Tensor};
 use crate::math::torus::UnsignedTorus;
 use crate::test_tools::{
     assert_delta_std_dev, assert_noise_distribution, random_ciphertext_count, random_lwe_dimension,
     random_uint_between,
+};
+use concrete_commons::dispersion::{DispersionParameter, LogStandardDev, Variance};
+use concrete_commons::numeric::{CastFrom, Numeric, SignedInteger};
+use concrete_commons::parameters::{
+    CiphertextCount, CleartextCount, DecompositionBaseLog, DecompositionLevelCount, LweDimension,
+    PlaintextCount,
 };
 
 fn test_keyswitch<T: UnsignedTorus + RandomGenerable<UniformMsb> + npe::LWE>() {
