+++ concrete-core/src/crypto/secret/generators/secret.rs
@@ -1,7 +1,7 @@
 use crate::math::random::{Gaussian, RandomGenerable, RandomGenerator};
 use crate::math::tensor::Tensor;
 use crate::math::torus::UnsignedTorus;
-use concrete_commons::DispersionParameter;
+use concrete_commons::dispersion::DispersionParameter;
 
 /// A random number generator which can be used to generate secret keys.
 pub struct SecretRandomGenerator(RandomGenerator);
