+++ concrete-core/src/crypto/secret/lwe.rs
@@ -1,17 +1,17 @@
 use serde::{Deserialize, Serialize};
 
-use concrete_commons::{
-    BinaryKeyKind, DispersionParameter, GaussianKeyKind, KeyKind, Numeric, TernaryKeyKind,
-    UniformKeyKind,
-};
-
 use crate::crypto::encoding::{Plaintext, PlaintextList};
 use crate::crypto::lwe::{LweCiphertext, LweList};
 use crate::crypto::secret::generators::{EncryptionRandomGenerator, SecretRandomGenerator};
-use crate::crypto::LweDimension;
 use crate::math::random::{Gaussian, RandomGenerable};
 use crate::math::tensor::{AsMutSlice, AsMutTensor, AsRefSlice, AsRefTensor, IntoTensor, Tensor};
 use crate::math::torus::UnsignedTorus;
+use concrete_commons::dispersion::DispersionParameter;
+use concrete_commons::key_kinds::{
+    BinaryKeyKind, GaussianKeyKind, KeyKind, TernaryKeyKind, UniformKeyKind,
+};
+use concrete_commons::numeric::Numeric;
+use concrete_commons::parameters::LweDimension;
 use std::marker::PhantomData;
 
 /// A LWE secret key.
@@ -34,6 +34,7 @@ where
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::LweDimension;
     /// use concrete_core::crypto::secret::generators::SecretRandomGenerator;
     /// use concrete_core::crypto::secret::*;
     /// use concrete_core::crypto::*;
@@ -60,6 +61,7 @@ where
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::LweDimension;
     /// use concrete_core::crypto::secret::generators::SecretRandomGenerator;
     /// use concrete_core::crypto::secret::*;
     /// use concrete_core::crypto::*;
@@ -87,6 +89,7 @@ where
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::LweDimension;
     /// use concrete_core::crypto::secret::generators::SecretRandomGenerator;
     /// use concrete_core::crypto::secret::*;
     /// use concrete_core::crypto::*;
@@ -113,6 +116,7 @@ where
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::LweDimension;
     /// use concrete_core::crypto::secret::generators::SecretRandomGenerator;
     /// use concrete_core::crypto::secret::*;
     /// use concrete_core::crypto::*;
@@ -141,6 +145,7 @@ impl<Cont> LweSecretKey<BinaryKeyKind, Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::LweDimension;
     /// use concrete_core::crypto::secret::*;
     /// use concrete_core::crypto::*;
     /// let secret_key = LweSecretKey::binary_from_container(vec![true; 256]);
@@ -169,6 +174,7 @@ impl<Cont> LweSecretKey<TernaryKeyKind, Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::LweDimension;
     /// use concrete_core::crypto::secret::*;
     /// use concrete_core::crypto::*;
     /// let secret_key = LweSecretKey::ternary_from_container(vec![true; 256]);
@@ -197,6 +203,7 @@ impl<Cont> LweSecretKey<GaussianKeyKind, Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::LweDimension;
     /// use concrete_core::crypto::secret::*;
     /// use concrete_core::crypto::*;
     /// let secret_key = LweSecretKey::gaussian_from_container(vec![true; 256]);
@@ -225,6 +232,7 @@ impl<Cont> LweSecretKey<UniformKeyKind, Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::LweDimension;
     /// use concrete_core::crypto::secret::*;
     /// use concrete_core::crypto::*;
     /// let secret_key = LweSecretKey::uniform_from_container(vec![true; 256]);
@@ -250,6 +258,7 @@ where
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::LweDimension;
     /// use concrete_core::crypto::secret::*;
     /// use concrete_core::crypto::*;
     /// let secret_key = LweSecretKey::binary_from_container(vec![true; 256]);
@@ -267,8 +276,8 @@ where
     /// # Example
     ///
     /// ```rust
-    /// use concrete_commons::LogStandardDev;
-    ///
+    /// use concrete_commons::dispersion::LogStandardDev;
+    /// use concrete_commons::parameters::{LweDimension, LweSize};
     /// use concrete_core::crypto::encoding::*;
     /// use concrete_core::crypto::lwe::*;
     /// use concrete_core::crypto::secret::generators::{
@@ -330,8 +339,10 @@ where
     /// # Example
     ///
     /// ```rust
-    /// use concrete_commons::LogStandardDev;
-    ///
+    /// use concrete_commons::dispersion::LogStandardDev;
+    /// use concrete_commons::parameters::{
+    ///     CiphertextCount, CleartextCount, LweDimension, LweSize, PlaintextCount,
+    /// };
     /// use concrete_core::crypto::encoding::*;
     /// use concrete_core::crypto::lwe::*;
     /// use concrete_core::crypto::secret::generators::{
@@ -388,7 +399,7 @@ where
             "Lwe cipher list size and encoded list size are not compatible"
         );
         for (mut cipher, message) in output.ciphertext_iter_mut().zip(encoded.plaintext_iter()) {
-            self.encrypt_lwe(&mut cipher, message, noise_parameters.clone(), generator);
+            self.encrypt_lwe(&mut cipher, message, noise_parameters, generator);
         }
     }
 
@@ -397,8 +408,8 @@ where
     /// # Example
     ///
     /// ```rust
-    /// use concrete_commons::LogStandardDev;
-    ///
+    /// use concrete_commons::dispersion::LogStandardDev;
+    /// use concrete_commons::parameters::{LweDimension, LweSize};
     /// use concrete_core::crypto::encoding::*;
     /// use concrete_core::crypto::lwe::*;
     /// use concrete_core::crypto::secret::generators::{
@@ -462,8 +473,10 @@ where
     /// # Example
     ///
     /// ```rust
-    /// use concrete_commons::LogStandardDev;
-    ///
+    /// use concrete_commons::dispersion::LogStandardDev;
+    /// use concrete_commons::parameters::{
+    ///     CiphertextCount, CleartextCount, LweDimension, LweSize, PlaintextCount,
+    /// };
     /// use concrete_core::crypto::encoding::*;
     /// use concrete_core::crypto::lwe::*;
     /// use concrete_core::crypto::secret::generators::{
@@ -526,7 +539,7 @@ where
             "Lwe cipher list size and encoded list size are not compatible"
         );
         for (mut cipher, message) in output.ciphertext_iter_mut().zip(encoded.plaintext_iter()) {
-            self.trivial_encrypt_lwe(&mut cipher, message, noise_parameters.clone(), generator);
+            self.trivial_encrypt_lwe(&mut cipher, message, noise_parameters, generator);
         }
     }
