+++ concrete-core/src/lib.rs
@@ -25,14 +25,14 @@
 //! ```
 //! // This examples shows how to multiply a secret value by a public one homomorphically. First
 //! // we import the proper symbols:
-//! use concrete_commons::LogStandardDev;
+//! use concrete_commons::dispersion::LogStandardDev;
+//! use concrete_commons::parameters::LweDimension;
 //! use concrete_core::crypto::encoding::{Cleartext, Encoder, Plaintext, RealEncoder};
 //! use concrete_core::crypto::lwe::LweCiphertext;
 //! use concrete_core::crypto::secret::generators::{
 //!     EncryptionRandomGenerator, SecretRandomGenerator,
 //! };
 //! use concrete_core::crypto::secret::LweSecretKey;
-//! use concrete_core::crypto::LweDimension;
 //!
 //! // We initialize a prng that will be used to generate secret keys.
 //! let mut secret_generator = SecretRandomGenerator::new(None);
@@ -122,15 +122,15 @@ pub mod utils;
 pub mod test_tools {
     use rand::Rng;
 
-    use concrete_commons::DispersionParameter;
-
-    use crate::crypto::{CiphertextCount, GlweDimension, LweDimension, PlaintextCount};
-    use crate::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount};
-    use crate::math::polynomial::PolynomialSize;
     use crate::math::random::{RandomGenerable, RandomGenerator, Uniform};
     use crate::math::tensor::{AsRefSlice, AsRefTensor};
     use crate::math::torus::UnsignedTorus;
-    use concrete_commons::UnsignedInteger;
+    use concrete_commons::dispersion::DispersionParameter;
+    use concrete_commons::numeric::UnsignedInteger;
+    use concrete_commons::parameters::{
+        CiphertextCount, DecompositionBaseLog, DecompositionLevelCount, GlweDimension,
+        LweDimension, PlaintextCount, PolynomialSize,
+    };
 
     fn modular_distance<T: UnsignedInteger>(first: T, other: T) -> T {
         let d0 = first.wrapping_sub(other);
