+++ concrete-core/src/math/decomposition/tests.rs
@@ -1,9 +1,10 @@
-use crate::math::decomposition::{DecompositionBaseLog, DecompositionLevelCount, SignedDecomposer};
+use crate::math::decomposition::SignedDecomposer;
 use crate::math::random::{RandomGenerable, Uniform};
 use crate::math::tensor::Tensor;
 use crate::math::torus::UnsignedTorus;
 use crate::test_tools::{any_uint, any_usize, random_usize_between};
-use concrete_commons::{Numeric, SignedInteger, UnsignedInteger};
+use concrete_commons::numeric::{Numeric, SignedInteger, UnsignedInteger};
+use concrete_commons::parameters::{DecompositionBaseLog, DecompositionLevelCount};
 use std::fmt::Debug;
 
 // Returns a random decomposition valid for the size of the T type.
