+++ concrete-core/src/math/fft/polynomial.rs
@@ -1,11 +1,11 @@
 use concrete_fftw::array::AlignedVec;
 use serde::{Deserialize, Serialize};
 
-use crate::math::polynomial::PolynomialSize;
 use crate::math::tensor::{AsMutSlice, AsMutTensor, AsRefSlice, AsRefTensor, Tensor};
 use crate::{ck_dim_eq, tensor_traits, zip, zip_args};
 
 use super::Complex64;
+use concrete_commons::parameters::PolynomialSize;
 
 /// A polynomial in the fourier domain.
 ///
@@ -23,8 +23,8 @@ impl FourierPolynomial<AlignedVec<Complex64>> {
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, FourierPolynomial};
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let fourier_poly = FourierPolynomial::allocate(Complex64::new(0., 0.), PolynomialSize(128));
     /// assert_eq!(fourier_poly.polynomial_size(), PolynomialSize(128));
     /// ```
@@ -41,8 +41,8 @@ impl<Cont> FourierPolynomial<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{AlignedVec, Complex64, FourierPolynomial};
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let mut alvec: AlignedVec<Complex64> = AlignedVec::new(128);
     /// let fourier_poly = FourierPolynomial::from_container(alvec.as_slice_mut());
     /// assert_eq!(fourier_poly.polynomial_size(), PolynomialSize(128));
@@ -62,8 +62,8 @@ impl<Cont> FourierPolynomial<Cont> {
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, FourierPolynomial};
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let fourier_poly = FourierPolynomial::allocate(Complex64::new(0., 0.), PolynomialSize(128));
     /// assert_eq!(fourier_poly.polynomial_size(), PolynomialSize(128));
     /// ```
@@ -83,8 +83,8 @@ impl<Cont> FourierPolynomial<Cont> {
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, FourierPolynomial};
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let fourier_poly = FourierPolynomial::allocate(Complex64::new(0., 0.), PolynomialSize(128));
     /// for coef in fourier_poly.coefficient_iter() {
     ///     assert_eq!(*coef, Complex64::new(0., 0.));
@@ -107,8 +107,8 @@ impl<Cont> FourierPolynomial<Cont> {
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, FourierPolynomial};
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// use concrete_core::math::tensor::AsRefTensor;
     /// let mut fourier_poly = FourierPolynomial::allocate(Complex64::new(0., 0.), PolynomialSize(128));
     /// for mut coef in fourier_poly.coefficient_iter_mut() {
@@ -136,8 +136,8 @@ impl<Cont> FourierPolynomial<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, FourierPolynomial};
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let mut fpoly1 = FourierPolynomial::allocate(Complex64::new(1., 2.), PolynomialSize(128));
     /// let fpoly2 = FourierPolynomial::allocate(Complex64::new(3., 4.), PolynomialSize(128));
     /// let fpoly3 = FourierPolynomial::allocate(Complex64::new(5., 6.), PolynomialSize(128));
@@ -178,8 +178,8 @@ impl<Cont> FourierPolynomial<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, FourierPolynomial};
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let mut fpoly1 = FourierPolynomial::allocate(Complex64::new(1., 2.), PolynomialSize(128));
     /// let fpoly2 = FourierPolynomial::allocate(Complex64::new(3., 4.), PolynomialSize(128));
     /// let fpoly3 = FourierPolynomial::allocate(Complex64::new(5., 6.), PolynomialSize(128));
@@ -232,8 +232,8 @@ impl<Cont> FourierPolynomial<Cont> {
     /// # Example
     ///
     /// ```rust
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, FourierPolynomial};
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// macro_rules! new_poly {
     ///     ($name: ident, $re: literal, $im: literal) => {
     ///         let mut $name =
