+++ concrete-core/src/math/fft/tests.rs
@@ -1,10 +1,10 @@
-use concrete_commons::*;
-
 use crate::math::fft::twiddles::{BackwardCorrector, ForwardCorrector};
 use crate::math::fft::{Complex64, Fft, FourierPolynomial, SerializableComplex64};
-use crate::math::polynomial::{Polynomial, PolynomialSize};
+use crate::math::polynomial::Polynomial;
 use crate::math::random::RandomGenerator;
 use crate::math::tensor::{AsMutTensor, AsRefTensor};
+use concrete_commons::numeric::Numeric;
+use concrete_commons::parameters::PolynomialSize;
 use concrete_fftw::array::AlignedVec;
 use serde_test::{assert_tokens, Token};
