+++ concrete-core/src/math/fft/transform.rs
@@ -3,15 +3,17 @@ use std::slice;
 use concrete_fftw::array::AlignedVec;
 use concrete_fftw::types::c64;
 
+use concrete_commons::numeric::{CastInto, SignedInteger, UnsignedInteger};
+use concrete_commons::parameters::PolynomialSize;
+
 use crate::math::fft::twiddles::{BackwardCorrector, ForwardCorrector};
-use crate::math::polynomial::{Polynomial, PolynomialSize};
+use crate::math::polynomial::Polynomial;
 use crate::math::tensor::{AsMutSlice, AsMutTensor, AsRefSlice, AsRefTensor};
 use crate::math::torus::UnsignedTorus;
 use crate::{ck_dim_eq, zip};
 
 use super::{Complex64, Correctors, FourierPolynomial};
 use crate::math::fft::plan::Plans;
-use concrete_commons::{CastInto, SignedInteger, UnsignedInteger};
 use std::cell::RefCell;
 
 /// A fast fourier transformer.
@@ -31,8 +33,8 @@ impl Fft {
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::Fft;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let fft = Fft::new(PolynomialSize(256));
     /// assert_eq!(fft.polynomial_size(), PolynomialSize(256));
     /// ```
@@ -60,8 +62,8 @@ impl Fft {
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::Fft;
-    /// use concrete_core::math::polynomial::PolynomialSize;
     /// let fft = Fft::new(PolynomialSize(256));
     /// assert_eq!(fft.polynomial_size(), PolynomialSize(256));
     /// ```
@@ -81,8 +83,9 @@ impl Fft {
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, Fft, FourierPolynomial};
-    /// use concrete_core::math::polynomial::{Polynomial, PolynomialSize};
+    /// use concrete_core::math::polynomial::Polynomial;
     /// use concrete_core::math::random::RandomGenerator;
     /// use concrete_core::math::tensor::AsRefTensor;
     /// use concrete_core::math::torus::UnsignedTorus;
@@ -118,8 +121,9 @@ impl Fft {
     /// # Example
     ///
     /// ```
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, Fft, FourierPolynomial};
-    /// use concrete_core::math::polynomial::{Polynomial, PolynomialSize};
+    /// use concrete_core::math::polynomial::Polynomial;
     /// use concrete_core::math::random::RandomGenerator;
     /// use concrete_core::math::tensor::AsRefTensor;
     /// use concrete_core::math::torus::UnsignedTorus;
@@ -194,9 +198,10 @@ impl Fft {
     /// # Example
     ///
     /// ```
-    /// use concrete_commons::UnsignedInteger;
+    /// use concrete_commons::numeric::UnsignedInteger;
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, Fft, FourierPolynomial};
-    /// use concrete_core::math::polynomial::{Polynomial, PolynomialSize};
+    /// use concrete_core::math::polynomial::Polynomial;
     /// use concrete_core::math::random::RandomGenerator;
     /// use concrete_core::math::tensor::AsRefTensor;
     /// let mut generator = RandomGenerator::new(None);
@@ -231,9 +236,10 @@ impl Fft {
     /// # Example
     ///
     /// ```
-    /// use concrete_commons::UnsignedInteger;
+    /// use concrete_commons::numeric::UnsignedInteger;
+    /// use concrete_commons::parameters::PolynomialSize;
     /// use concrete_core::math::fft::{Complex64, Fft, FourierPolynomial};
-    /// use concrete_core::math::polynomial::{Polynomial, PolynomialSize};
+    /// use concrete_core::math::polynomial::Polynomial;
     /// use concrete_core::math::random::RandomGenerator;
     /// use concrete_core::math::tensor::AsRefTensor;
     /// let mut generator = RandomGenerator::new(None);
