+++ concrete-core/src/math/polynomial/mod.rs
@@ -8,8 +8,6 @@
 //! + [`PolynomialList`], which represent a set of polynomials with the same degree, on which
 //! operations can be performed.
 
-use serde::{Deserialize, Serialize};
-
 pub use list::*;
 pub use monomial::*;
 pub use polynomial::*;
@@ -27,15 +25,3 @@ mod polynomial;
 /// Assuming a monomial $aX^N$, this returns the $N$ value.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct MonomialDegree(pub usize);
-
-/// The number of coefficients of a polynomial.
-///
-/// Assuming a polynomial $a_0 + a_1X + /dots + a_nX^N$, this returns $N+1$.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PolynomialSize(pub usize);
-
-/// The number of polynomials in a polynomial list.
-///
-/// Assuming a polynomial list, this return the number of polynomials.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct PolynomialCount(pub usize);
