+++ concrete-core/src/math/polynomial/tests.rs
@@ -1,7 +1,10 @@
-use crate::math::polynomial::{MonomialDegree, Polynomial, PolynomialSize};
+use rand::Rng;
+
+use concrete_commons::parameters::PolynomialSize;
+
+use crate::math::polynomial::{MonomialDegree, Polynomial};
 use crate::math::random::RandomGenerator;
 use crate::math::torus::UnsignedTorus;
-use rand::Rng;
 
 fn test_multiply_divide_unit_monomial<T: UnsignedTorus>() {
     //! tests if multiply_by_monomial and divide_by_monomial cancel each other
