+++ concrete-core/src/math/random/uniform_msb.rs
@@ -1,4 +1,5 @@
 use super::*;
+use concrete_commons::numeric::Numeric;
 
 /// A distribution type representing random sampling for unsigned integer types, where the `n`
 /// most significant bits are sampled randomly in `[0, 2^n[`.
