+++ concrete-core/src/math/random/uniform_with_zeros.rs
@@ -1,4 +1,5 @@
 use super::*;
+use concrete_commons::numeric::Numeric;
 
 /// A distribution type that samples a uniform value with probability `1 - prob_zero`, and a zero
 /// value with probaibility `prob_zero`.
