+++ concrete-core/src/math/torus/mod.rs
@@ -13,9 +13,8 @@
 //! floating point representation.
 
 use crate::math::random::{Gaussian, RandomGenerable, Uniform, UniformBinary, UniformTernary};
-use concrete_commons::{
-    CastFrom, CastInto, FloatingPoint, LogStandardDev, Numeric, UnsignedInteger,
-};
+use concrete_commons::dispersion::LogStandardDev;
+use concrete_commons::numeric::{CastFrom, CastInto, FloatingPoint, Numeric, UnsignedInteger};
 use std::fmt::{Debug, Display};
 
 /// A trait that converts a torus element in unsigned integer representation to the closest
